Snap: HexWorld.java, Hallway.java, HallwayUnit.java, RoomUnit.java, Position.java, RandomWorld.java, Counter.java, BoringWorldDemo.java

+++ proj3/byow/Core/RandomWorld.java
@@ -92,7 +92,7 @@ public class RandomWorld {
      */
     private static void generateWorld(long seed, TETile[][] world, Position focus) {
         RoomUnit r = randomlyGeneration(seed, world, focus);
-        int complexity = 0;
+        double complexity = 0;
         while (complexity < 0.5) {
             seed = RANDOM.nextInt((int) seed);
             r = randomlyGeneration(seed, world, (r.getExits()[0]));
